requirements.txt
telethon 
tweepy
gspread
python-dotenv
networkx
matplotlib